+++ lib/scalyr/common/util.rb
@@ -4,7 +4,7 @@ module Scalyr; module Common; module Util;
 # Flattens a hash or array, returning a hash where keys are a delimiter-separated string concatenation of all
 # nested keys.  Returned keys are always strings.  If a non-hash or array is provided, raises TypeError.
 # Please see rspec util_spec.rb for expected behavior.
-def self.flatten(obj, delimiter='_')
+def self.flatten(obj, delimiter='_', flatten_arrays=true)
 
   # base case is input object is not enumerable, in which case simply return it
   if !obj.respond_to?(:each)
@@ -19,8 +19,8 @@ def self.flatten(obj, delimiter='_')
 
     # input object is a hash
     obj.each do |key, value|
-      if value.respond_to?(:each)
-        flatten(value).each do |subkey, subvalue|
+      if (flatten_arrays and value.respond_to?(:each)) or value.respond_to?(:has_key?)
+        flatten(value, delimiter, flatten_arrays).each do |subkey, subvalue|
           result["#{key}#{delimiter}#{subkey}"] = subvalue
         end
       else
@@ -28,18 +28,23 @@ def self.flatten(obj, delimiter='_')
       end
     end
 
-  else
+  elsif flatten_arrays
 
     # input object is an array or set
     obj.each_with_index do |value, index|
       if value.respond_to?(:each)
-        flatten(value).each do |subkey, subvalue|
+        flatten(value, delimiter, flatten_arrays).each do |subkey, subvalue|
           result["#{index}#{delimiter}#{subkey}"] = subvalue
         end
       else
         result["#{index}"] = value
       end
     end
+
+  else
+
+    result = obj
+
   end
 
   return result
@@ -50,6 +55,27 @@ def self.truncate(content, max)
     return "#{content[0...(max-3)]}..."
   end
   return content
+end
+
+def self.convert_bignums(obj)
+  if obj.respond_to?(:has_key?) and obj.respond_to?(:each)
+    # input object is a hash
+    obj.each do |key, value|
+      obj[key] = convert_bignums(value)
+    end
+
+  elsif obj.respond_to?(:each)
+    # input object is an array or set
+    obj.each_with_index do |value, index|
+      obj[index] = convert_bignums(value)
+    end
+
+  elsif obj.is_a? Bignum
+    return obj.to_s
+
+  else
+    return obj
+  end
 end
 
 end; end; end;