+++ lib/scalyr/constants.rb
@@ -1,2 +1,2 @@
 # encoding: utf-8
-PLUGIN_VERSION = "v0.1.16.beta"
+PLUGIN_VERSION = "v0.1.17.beta"