+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## TBD
 - Add metrics for successfully sent and failed logstash events, and retries.
+- Make array flattening optional during nested value flattening with the `flatten_nested_arrays` configuration option.
+
+## 0.1.17.beta
+- Catch errors relating to Bignum conversions present in the ``json`` library and manually convert to string as
+a workaround.
 
 ## 0.1.16.beta
 - Fix race condition in ``register()`` method.